shift instruction selection

+++ go/amd64/instructions/type_conversion.go
@@ -256,7 +256,7 @@ func convertUint64ToFloat(
 	copyGeneral(instructions, 8, scratch, src)
 
 	// <scratch> = <scratch> / 2
-	shrIntImmediate(instructions, ir.Uint64, scratch, 1)
+	shrIntImmediate(instructions, ir.Uint64, scratch, uint8(1))
 
 	// <src> = <src> % 2
 	andIntImmediate(instructions, ir.Uint32, src, uint32(1))
